requirements.txt
Django
scikit-learn
pandas
numpy
django-cors-headers
selenium
webdriver-manager
whois
beautifulsoup4
requests
joblib
re
google-generativeai
dotenv
nltk
pyaxmlparser